Decorators moved to Helpers. Some trivial updates.
